shared form for prints
