+++ frontend/app/pages/admin/prints/index.vue
@@ -17,7 +17,6 @@ import type { PrintItemView, PaginatedResponse } from "@repo/common";
 import { usePrintApi } from "../../../composables/media/usePrintApi";
 import { usePrintView } from "../../../composables/media/usePrintView";
 import PrintsToolbar from "../../../components/prints/PrintsToolbar.vue";
-import FileList from "../../../components/admin/prints/FileList.vue";
 
 const { t, locale } = useI18n();
 const { getAll, remove } = usePrintApi();
@@ -122,7 +121,7 @@ onMounted(() => {
 
   <!-- File list -->
   <div class="container mx-auto px-4 max-w-5xl py-6">
-    <FileList
+    <AdminPrintsFileList
       :category="activeFilter"
       :files="prints"
       :total-pages="totalPages"
